Discriminate physical memory regions by type.
The physical memory allocator is a number of separate buddy
allocators which are all given separate memory regions with varying
properties. You may call mm_physical_alloc and specify a type of
region (DRAM, NVRAM, <4GB, etc), and the allocators will try to
fulfill the request, if possible.

The arch-dependent code is responsible for providing the region
information to the memory manager upon start-up, and must be able
to do so without allocating additional memory outside of the kernel's
image.