Allow updating permissions on EPT entries during objspace fault.
Previously, the EPT code would simply refuse to change a mapping.
However, this is too strict. It's useful to map, upon first access,
an object with its default permissions if those are sufficient, since
there's a good chance that will be enough for the execution and we can
avoid the expensive operation of accessing security contexts. However,
if later we need more access rights, and the security context has
them, we'll need to update the EPT entry.

So now you can pass OBJSPACE_SET_FLAGS, which will allow updating
permission bits IF AND ONLY IF the supplied physical address matches
the one already present in the EPT entry.