Add .gates section to linker script.
Allows generation of gates (via twz/gates.h) for programs. The gates
section goes at the start of the text phdr at a specific location. We
put it there via the extra readonly section parameter for the emulation
in ld; this should be fine, but it _might_ be fragile depending on if
binutils moves this around. We hack it a bit by adding the . =
ALIGN(0x200) specifier at the beginning, even though the param is
supposed to only be a section specifier for the ldscript. Works, though.