Implement opin and octl.
The opin syscall is necessary for drivers to submit io memory
addresses to devices. It allows drivers to require that the kernel
does not evict the object logical slot for an object without some
arbitration method (not yet defined).

The octl syscall is necessary for modifying certain properties of
objects. For now, these properties are volatile (they are not kept
across reboots). Currently, we implement OCO_CACHE_MODE, which lets
an application specify the caching policy of a data range of an
object. We need this for drivers again, since they often need to
specify a less-than-WB cache policy on pages used for command queues
etc.

Secondary: use these interfaces in NVMe driver, and make NVMe driver
track dstride better.