us-interrupts support, nvme identify cmd gen, queues.
Implement basic cmd building and submission for nvme.

Support requesting interrupt vectors from userspace via the kaction
method. We'll need this for userspace drivers. However, the API is
definitely not finalized. It remains to be seen if we need kernel-side
support for some more complex devices, but for some devices I think we
should be okay.