Remove thread from all sp's lists on wakeup.
When a thread wakes up, it either does so by itself or by another thread
waking it up. When it wakes up by itself, it can easily just remove
itself from all syncpoints. But when another thread wakes it up, the
waking thread now traverses all the sleep_entries in a thread and
removes the thread from the other syncpoints lists. This is important
because otherwise, a thread could be kept on a list too long, leading to
silent corruption.

Important: the thread is removed from all lists BEFORE it is woken. And
we need to hold the thread's lock while doing so.