Enable thread support in the hosted cross compiler.
Typically, the threading system doesn't actually require a lot of
compiler support, but the stdc++ library checks to see if it should
allow C++11 threading based on if the -pthread option works. So to get
the C++11 threading support, we'll need the -pthread option to work.

To get this to work, we need to specify the options file for gcc/config
to include the -pthread driver (see gcc/config/twizzler.opt), and add
the tfile for threading in libgcc, along with enabling the posix thread
file for libgcc.

This commit add these features to the patch for gcc 9.2.0 to make the
cross compiler.