Use x2APIC mode instead of xAPIC mode.
This has several benefits: APIC register access is done via MSRs, which
is more efficient, and doesn't require complex memory management for
setting it up. It also enables scalability beyond 255 processors, and
better IPI sending methods.

However, Intel does not document which processor models support it, as
it is "optional". We're requiring it. Justification is that most
"reasonably modern" intel chips will support it, and while AMD does not
support it, we don't support AMD anyway yet.

The reason for this change now is that the new Xeon Scalable chips use
x2APIC and it seems that some manufacturers enable in the BIOS during
start. Disabling the x2APIC mode is almost as complicated as just
supporting the thing in the first place, and I've been meaning to get it
to work anyway.