Release pages, alloc from buddy first.
So, this commit releases a lot of pages that are allocated during normal
operation. It also changes the policy of allocation to try to allocate
from the buddy allocator before trying to allocate from the alloc-arena
(that contains pages to-be-freed into the buddy allocator). This is
because allocations from the arena almost certainly require an
object-space fault to setup the page. But allocations from the buddy
allocator do not.