Implement select(2) for Twix.
This adds to the twzio API a poll function that both, 1) constructs a
struct event that can be waited on correctly to wait for a given event,
and 2) returns if that event is currently triggered.

Implements support for poll in bstream and pty.

Implements a select(2) function that uses the poll part of twzio to wait
for file descriptors.

Add vim port (currently partially works).