Support 32-bit words for thread_sync.
This is necessary to support the futex Linux syscall, which uses 32-bit
words. The kernel derives syncpoints independent of the word size the
thread_sync call is using, so there may be odd behavior if programs are
sleeping on the same word with different word sizes. This should not
matter in a well-type program :)