Fix sysreturn invalid rcx bug.
x86 has this thing where if rcx is a non-canonical address when
returning from syscall, it will trigger an exception _from the kernel_.
But because, in the exception handler, we do a swapgs based on if its a
user exception or not, and when we return from syscall we do a swapgs
before executing the return instruction, if rcx is non-canonical, we'll
enter the kernel's exception handler with RPL 0, with a swapped gs like
we're in userspace. This triggers the kernel to _not_ swapgs, and run
with user's gs. This is bad, because we rely on gs to locate ourselves
on exception entry.

Check if rcx is canonical when returning from syscall. If its not, raise
a fault. Add a check in arch_thread_resume to see if the thread has
exited right before doing the final context switch. If it has, just
reschedule.